+++ src/main/java/skylands/event/SkylandsEvents.java
@@ -39,8 +39,8 @@ public class SkylandsEvents {
 				}
 			}
 		});
-		ServerLifecycleEvents.SERVER_STARTING.register(ServerStartEvent::onStart);
-		ServerLifecycleEvents.SERVER_STARTED.register(ServerStartedEvent::onStarted);
+		ServerLifecycleEvents.SERVER_STARTING.register(ServerStartEvent::onStarting);
+		ServerLifecycleEvents.SERVER_STARTED.register(ServerStartEvent::onStarted);
 		ServerTickEvents.END_SERVER_TICK.register(ServerTickEvent::onTick);
 		ServerPlayConnectionEvents.JOIN.register((handler, sender, server) -> PlayerConnectEvent.onJoin(server, handler.player));
 		ServerPlayConnectionEvents.DISCONNECT.register((handler, server) -> PlayerConnectEvent.onLeave(server, handler.player));
