+++ src/main/resources/skylands.mixins.json
@@ -6,8 +6,8 @@
   "mixins": [
     "block.FireBlockMixin",
     "entity.PlayerEntityMixin",
-    "gen.NoiseChunkGeneratorMixin",
-    "gen.StructureAccessorAccessor",
+    "world.gen.NoiseChunkGeneratorMixin",
+    "world.gen.StructureAccessorAccessor",
     "world.ServerWorldMixin",
     "world.protection.ArmorStandMixin",
     "world.protection.AxeItemMixin",
